+++ src/elements/core/interfaces/overlay-close-details.ts
@@ -1,3 +1,4 @@
+/** @deprecated */
 export type SbbOverlayCloseEventDetails = {
   returnValue?: any;
   closeTarget?: HTMLElement;