+++ src/elements/divider/__snapshots__/divider.snapshot.spec.snap.js
@@ -7,35 +7,17 @@ snapshots["sbb-divider renders DOM"] =
 `;
 /* end snapshot sbb-divider renders DOM */
 
-snapshots["sbb-divider renders Shadow DOM"] = 
-`<div class="sbb-divider">
-</div>
-`;
-/* end snapshot sbb-divider renders Shadow DOM */
-
 snapshots["sbb-divider renders horizontal DOM"] = 
 `<sbb-divider orientation="horizontal">
 </sbb-divider>
 `;
 /* end snapshot sbb-divider renders horizontal DOM */
 
-snapshots["sbb-divider renders horizontal Shadow DOM"] = 
-`<div class="sbb-divider">
-</div>
-`;
-/* end snapshot sbb-divider renders horizontal Shadow DOM */
-
 snapshots["sbb-divider renders vertical DOM"] = 
 `<sbb-divider orientation="vertical">
 </sbb-divider>
 `;
 /* end snapshot sbb-divider renders vertical DOM */
-
-snapshots["sbb-divider renders vertical Shadow DOM"] = 
-`<div class="sbb-divider">
-</div>
-`;
-/* end snapshot sbb-divider renders vertical Shadow DOM */
 
 snapshots["sbb-divider renders A11y tree Chrome"] = 
 `<p>
@@ -56,4 +38,22 @@ snapshots["sbb-divider renders A11y tree Firefox"] =
 </p>
 `;
 /* end snapshot sbb-divider renders A11y tree Firefox */
+
+snapshots["sbb-divider renders Shadow DOM"] = 
+`<div class="sbb-divider">
+</div>
+`;
+/* end snapshot sbb-divider renders Shadow DOM */
+
+snapshots["sbb-divider renders horizontal Shadow DOM"] = 
+`<div class="sbb-divider">
+</div>
+`;
+/* end snapshot sbb-divider renders horizontal Shadow DOM */
+
+snapshots["sbb-divider renders vertical Shadow DOM"] = 
+`<div class="sbb-divider">
+</div>
+`;
+/* end snapshot sbb-divider renders vertical Shadow DOM */