+++ src/elements/tabs.ts
@@ -4,3 +4,4 @@
 export * from './tabs/tab.ts';
 export * from './tabs/tab-group.ts';
 export * from './tabs/tab-label.ts';
+export * from './tabs/tab-nav-bar/tab-nav-bar.component.ts';